+++ t/tap-basic.sh
@@ -24,8 +24,7 @@
 # Note that some of the features checked here are checked in other
 # test cases too, usually in a more thorough and detailed way.
 
-am_parallel_tests=yes
-. ./defs || Exit 1
+. test-init.sh
 
 fetch_tap_driver
 
@@ -81,7 +80,7 @@ $AUTOMAKE
 
 # Basilar usage and testsuite progress output.
 
-$MAKE check >stdout && { cat stdout; Exit 1; }
+$MAKE check >stdout && { cat stdout; exit 1; }
 cat stdout
 
 count_test_results total=20 pass=4 fail=4 xpass=4 xfail=4 skip=4 error=0
@@ -128,37 +127,37 @@ Bail out!
 ok 1
 END
 
-TESTS=bail.test $MAKE -e check >stdout && { cat stdout; Exit 1; }
+TESTS=bail.test $MAKE -e check >stdout && { cat stdout; exit 1; }
 cat stdout
 
 count_test_results total=1 pass=0 fail=0 xpass=0 xfail=0 skip=0 error=1
 
-test ! -f success.log
+test ! -e success.log
 test -f bail.log
 test -f test-suite.log
 
 grep '^ERROR: bail\.test - Bail out!' stdout
-grep '^PASS:' stdout && Exit 1
-test `$FGREP -c ': bail.test' stdout` -eq 1
-$FGREP 'success.test' stdout && Exit 1
+grep '^PASS:' stdout && exit 1
+test $($FGREP -c ': bail.test' stdout) -eq 1
+$FGREP 'success.test' stdout && exit 1
 
 # Override TEST_LOGS from the command line, making it point to a test
 # (ok.test) that has to be generated at make time.
 
 rm -f *.log *.test
 
-TEST_LOGS=ok.log $MAKE -e check >stdout || { cat stdout; Exit 1; }
+TEST_LOGS=ok.log $MAKE -e check >stdout || { cat stdout; exit 1; }
 cat stdout
 
 count_test_results total=3 pass=1 fail=0 xpass=0 xfail=1 skip=1 error=0
 
 test -f ok.test
 test -f ok.log
-test ! -f success.log
-test ! -f bail.log
+test ! -e success.log
+test ! -e bail.log
 test -f test-suite.log
 
-$EGREP '(bail|success)\.test' stdout && Exit 1
+$EGREP '(bail|success)\.test' stdout && exit 1
 
 cat > exp << 'END'
 PASS: ok.test 1
